# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ikigai Compass API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discovery/sessions:
    get:
      operationId: listDiscoverySessions
      tags: [discovery]
      summary: List discovery sessions
      responses:
        "200":
          description: List of discovery sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoverySession"
    post:
      operationId: createDiscoverySession
      tags: [discovery]
      summary: Create or resume discovery session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiscoverySessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverySession"
  /discovery/sessions/{id}:
    get:
      operationId: getDiscoverySession
      tags: [discovery]
      summary: Get discovery session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discovery session with answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverySessionWithAnswers"
    put:
      operationId: updateDiscoverySession
      tags: [discovery]
      summary: Update discovery session answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDiscoverySessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverySession"
  /discovery/sessions/{id}/complete:
    post:
      operationId: completeDiscoverySession
      tags: [discovery]
      summary: Mark discovery session complete and trigger synthesis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverySession"
  /discovery/reflections:
    get:
      operationId: listReflections
      tags: [discovery]
      summary: List reflection entries
      responses:
        "200":
          description: List of reflections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReflectionEntry"
    post:
      operationId: createReflection
      tags: [discovery]
      summary: Create reflection entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReflectionBody"
      responses:
        "201":
          description: Created reflection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReflectionEntry"
  /discovery/reflections/{id}:
    get:
      operationId: getReflection
      tags: [discovery]
      summary: Get reflection entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reflection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReflectionEntry"
    put:
      operationId: updateReflection
      tags: [discovery]
      summary: Update reflection entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReflectionBody"
      responses:
        "200":
          description: Updated reflection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReflectionEntry"
...
  /checkins:
    get:
      operationId: listCheckIns
      tags: [checkins]
      summary: List accountability check-ins
      responses:
        "200":
          description: List of check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckIn"
    post:
      operationId: createCheckIn
      tags: [checkins]
      summary: Submit accountability check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckInBody"
      responses:
        "201":
          description: Created check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckIn"
  /checkins/{id}:
    get:
      operationId: getCheckIn
      tags: [checkins]
      summary: Get check-in
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckIn"
  /coaching/sessions:
    get:
      operationId: listCoachingSessions
      tags: [coaching]
      summary: List coaching sessions
      responses:
        "200":
          description: List of coaching sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachingSession"
    post:
      operationId: createCoachingSession
      tags: [coaching]
      summary: Start a coaching session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCoachingSessionBody"
      responses:
        "201":
          description: Created coaching session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingSession"
  /coaching/sessions/{id}:
    get:
      operationId: getCoachingSession
      tags: [coaching]
      summary: Get coaching session with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coaching session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingSessionWithMessages"
...
  /checkins/stagnation/{planId}:
    get:
      tags: [Checkins]
      summary: Detect stagnation for a growth plan
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stagnation status for the plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StagnationResult"
        "404":
          description: Plan not found
...
  /coaching/sessions/{id}/messages:
    post:
      operationId: sendCoachingMessage
      tags: [Coaching]
      summary: Send a message in a coaching session (SSE stream)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendCoachingMessageBody"
      responses:
        "200":
          description: SSE stream of coaching response chunks
          content:
            text/event-stream: {}
        "404":
          description: Session not found
components:
  schemas:
...
    DiscoverySession:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [in_progress, completed]
        currentStep:
          type: integer
        totalSteps:
          type: integer
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - status
        - currentStep
        - totalSteps
        - createdAt
        - updatedAt
    DiscoverySessionWithAnswers:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [in_progress, completed]
        currentStep:
          type: integer
        totalSteps:
          type: integer
        answers:
          type: object
          additionalProperties: {}
        themes:
          type: array
          items:
            type: string
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - status
        - currentStep
        - totalSteps
        - answers
        - themes
        - createdAt
        - updatedAt
    CreateDiscoverySessionBody:
      type: object
      properties:
        mode:
          type: string
          enum: [quick, deep]
      required:
        - mode
    UpdateDiscoverySessionBody:
      type: object
      properties:
        currentStep:
          type: integer
        answers:
          type: object
          additionalProperties: {}
    ReflectionEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        sessionId:
          type: integer
        prompt:
          type: string
        content:
          type: string
        aiSummary:
          type: string
        themes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - prompt
        - content
        - themes
        - createdAt
        - updatedAt
    CreateReflectionBody:
      type: object
      properties:
        sessionId:
          type: integer
        prompt:
          type: string
        content:
          type: string
      required:
        - prompt
        - content
    UpdateReflectionBody:
      type: object
      properties:
        content:
          type: string
...
    CheckIn:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        period:
          type: string
        plannedActions:
          type: string
        actualActions:
          type: string
        blockers:
          type: string
        energyLevel:
          type: integer
        satisfactionLevel:
          type: integer
        nextSteps:
          type: string
        aiSuggestions:
          type: string
        momentumScore:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - period
        - createdAt
    CreateCheckInBody:
      type: object
      properties:
        planId:
          type: integer
        period:
          type: string
        plannedActions:
          type: string
        actualActions:
          type: string
        blockers:
          type: string
        energyLevel:
          type: integer
        satisfactionLevel:
          type: integer
        nextSteps:
          type: string
      required:
        - period
    CoachingSession:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [general, obstacle, replan, reflection, celebration]
        title:
          type: string
        status:
          type: string
          enum: [active, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - status
        - createdAt
        - updatedAt
    CoachingSessionWithMessages:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [general, obstacle, replan, reflection, celebration]
        title:
          type: string
        status:
          type: string
          enum: [active, completed]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/CoachingMessage"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - status
        - messages
        - createdAt
        - updatedAt
    CoachingMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - role
        - content
        - createdAt
    CreateCoachingSessionBody:
      type: object
      properties:
        type:
          type: string
          enum: [general, obstacle, replan, reflection, celebration]
        title:
          type: string
        contextObstacleId:
          type: integer
        contextPlanId:
          type: integer
      required:
        - type
        - title
...
    StagnationResult:
      type: object
      properties:
        isStagnant:
          type: boolean
        daysSinceLastCheckin:
          type: integer
        thresholdDays:
          type: integer
        planId:
          type: integer
      required:
        - isStagnant
        - daysSinceLastCheckin
        - thresholdDays
...
    SendCoachingMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content